+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Getter `isNullOrEmpty` - returns `true` if the string is `null` or empty.
 * Getter `isNotNullNorEmpty` - returns `true` if the string is not `null` and not empty.
+* Method `firstToUpper()` - converts first character in the string to upper case.